Enhancements and bug fix
It is unnecessary to send two signals to the other albert instance.
One is enough to get the functionality.

Bug fix:
If albert was not terminated cleanly the file has still the invalid PID.
If albert starts and detects an invalid PID it will proceed but not update
the PID. So if a second instance is started the invalid PID will be detected
agail and still not updated and so multiple alberts will be running.
To fix this, invalid PIDs will simply be updated.