Adds single instance check
Uses QLocalServer to listen for new instances starting. If a new
instance successfully connects to the running instance, the new instance
quits and the existing instance will show the main window.

Future improvements could allow a remote app to control the running
instance, e.g send a command to show the settings or quit.

Fixes #23