Drop confusing minute timer implementation.
Scan interval timeout has been implemented by a timer with an interval
of 60 seconds. This introduced the need to sync multiple redundant
variables. Since INT allows to represent about 600 hrs in milliseconds
there is no need for this. Scan intervall timeout is now stored in the
timer intervall as product with 60000, which is the amount of
milliseconds of a minute.