Receive KeyPress events that otherwise be ignored
Whenever an application has a shortcut bound to a modifier (e.g. the
`Super` key), and the Albert launcher uses a shortcut based on the same
key, but as a modifier (e.g. `<Super>+Space`), Albert needs to register
to capture the KeyPress event that would otherwise be dealt with on the
other application and thus ignored by Albert.

This fix allows Albert to honor shortcuts that use modifiers bound to
other applications.