Fix crashes on Wayland by not loading the HotkeyManager
On Wayland, global hotkeys are often handled by the compositor instead. On sway
for instance, a config file is edited to launch specific commands when a hotkey
is entered.

On such compositors, it should be enough to show Albert by configuring the
compositor to launch "albert show" or "albert toggle" a hotkey manager in
Albert itself is redundant.