Move check for empty commandline to constructor of TermAction
Turns out that `ProcAction::activate` already checks for an empty
commandline. Additionally moving the check for an empty commandline to
the constructor changes the behaviour slightly as previously it was
possible for the `commandline_` variable to be filled with some string
even though it was completely empty and should not have been.