Fix #789. (#852)
Fix multiple terminals launch issue.

Closes https://github.com/albertlauncher/albert/issues/789.

The command to be executed by TermAction was calculated on every
activation. This calculation was not free of sideeffects and prepended
the terminalcommand on every run.
By moving these calculation to the constructor the problem has been
solved.

* Move check for empty commandline to constructor of TermAction

Turns out that `ProcAction::activate` already checks for an empty
commandline. Additionally moving the check for an empty commandline to
the constructor changes the behaviour slightly as previously it was
possible for the `commandline_` variable to be filled with some string
even though it was completely empty and should not have been.

* Always prepend current terminalCommand

It's possible for terminalCommand to change during the lifetime of a
TermAction and thus the full command line needs to be assembled every
time the action is activated.

* Remove unused member variables.