Add requisite OS
I couldn't tell if this project whether or not this project was for Linux, so after a bit of poking around, I learned that it is (and doesn't appear to be x-platform).  Added this critical detail to the README. :)

If it's compatible with other operating systems, it would be beneficial to so indicate in the README as well.

Thank you for albert!