Give enough time to connect to other instance.
When a first instance does not respond below 100 ms a second instance
assumes there is no other instance and the socket is a leftover of a
crash. This circumstance may happen when session managers autostart but
also restore albert from a former session at boot and the first instance
is simply to busy loading plugins.