Revert printing to logfile
Everything printed to cout or cerr will not be passed through the
messageHandler and as such not printed to the logfile. This makes this
logfile useless.

Redirecting cout using rdbuf is not an option either because C-style
printing is not covered.

There are system specific ways to tee the stdout/err fd, but thats to
cumbersome for now. Looks like we have to stick with shell solutions.