Change frontend interface design
Dynamic plugin/extension loading and unloading requires the core to
manage the lifetime of the queries to allow proper undloading of shared
objects.

Introduce session objects which manage queries by RAII.

Also slighly reduce the Query interface such that the actions are not
exposed to the frontend. Use strings and ids. Cache the created actions.

Internally serialize the triggered queries signals. Queued signals from
other threads may fire mulitple times which messes up the frontend state
machines. So we collect the results in the main thread using a buffer.