Update fallbacks when the widget is shown
This is a workaround such that FallbackHandlers dont need a signal for
the change of fallbacks. This is a bit dirty, but a cheap solution
maintenance and performance wise. The alternative would be to have a
signal in FallbackHandler and connect change the fallbacksmodel to
listen to it. That would be a lot of overhead for a very simple thing.