[API:add] Add dedicated empty query handling
Empty patterns should match everything. For global queries thats too
much. For triggered queries it is desired though. Since a lot of global
query handlers relay the handleTriggerQuery to handleGlobalQuery it is
not possible to have both. This introduces a dedicated function for
GlobalQueryHandlers that will be called on empty queries: