perf(globalqueryhandler): Instant response times in handleTriggerQuery
Let the default handleTriggerQuery implementation of the GlobalQueryHandler
perform a partial sort on the first bunch of items and add these before sorting
the remainder to avoid letting the user notice the lag the sort takes. This lag
can get quite huge, e.g. the docs with 10k items take 300ms to sort. With this
improvement response is always below 10ms.