feat(tray): Install the app tray icon on xdg
Currently the icon lookup order of ["xdg:albert-tray", ":app_tray_icon"] does
not use the tray icon but the installed albert icon. This happens due to the
icon lookup algorithm which truncates the last token and retries on failure.
This change installs the albert-tray icon such that the lookup prefers it over
the app icon.